Refactor index fetching logic by moving the fetchIndex function from enrichment.js to commerce.js. This change centralizes index data handling and improves code organization. Additionally, update loadCommerceLazy to ensure proper initialization sequence.
