+++ blocks/enrichment/enrichment.js
@@ -1,52 +1,6 @@
 import { readBlockConfig } from '../../scripts/aem.js';
-import { getSkuFromUrl } from '../../scripts/commerce.js';
+import { getSkuFromUrl, fetchIndex } from '../../scripts/commerce.js';
 import { loadFragment } from '../fragment/fragment.js';
-
-/**
- * Fetches and merges index data from multiple sources with intelligent caching.
- * @param {string} indexFile - The index file to fetch
- * @param {number} pageSize - The page size for pagination
- * @returns {Promise<Object>} A promise that resolves the index object
- */
-async function fetchIndex(indexFile, pageSize = 500) {
-  const handleIndex = async (offset) => {
-    const resp = await fetch(`/${indexFile}.json?limit=${pageSize}&offset=${offset}`);
-    const json = await resp.json();
-
-    const newIndex = {
-      complete: (json.limit + json.offset) === json.total,
-      offset: json.offset + pageSize,
-      promise: null,
-      data: [...window.index[indexFile].data, ...json.data],
-    };
-
-    return newIndex;
-  };
-
-  window.index = window.index || {};
-  window.index[indexFile] = window.index[indexFile] || {
-    data: [],
-    offset: 0,
-    complete: false,
-    promise: null,
-  };
-
-  // Return index if already loaded
-  if (window.index[indexFile].complete) {
-    return window.index[indexFile];
-  }
-
-  // Return promise if index is currently loading
-  if (window.index[indexFile].promise) {
-    return window.index[indexFile].promise;
-  }
-
-  window.index[indexFile].promise = handleIndex(window.index[indexFile].offset);
-  const newIndex = await (window.index[indexFile].promise);
-  window.index[indexFile] = newIndex;
-
-  return newIndex;
-}
 
 export default async function decorate(block) {
   const { type, position } = readBlockConfig(block);
